Extract checkbox, multiselect, select, radio
